Edit/delete buttons for Doctor/appointment
Adding delete/edit buttons for doctors and appointments

+++ src/main/java/com/example/demp/DoctorRepository.java
@@ -3,7 +3,7 @@ package com.example.demp;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface DoctorRepository extends JpaRepository<Doctor, Long> {
+public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
     @Query("SELECT u FROM Doctor u WHERE u.email = ?1")
     public Doctor findByEmail(String email);
 }
